Add files via upload
Heart disease remains one of the leading causes of morbidity and mortality worldwide, necessitating effective and early diagnosis to improve patient outcomes. This project focuses on the development and implementation of a predictive model for heart disease using machine learning techniques. The objective is to leverage historical patient data to identify patterns and risk factors that contribute to heart disease, thereby enabling early intervention and treatment.